requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
flake8==7.1.1
frozenlist==1.4.1
idna==3.8
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.1
tzdata==2024.1
yarl==1.11.1
drf-spectacular~=0.27.2
djangorestframework-simplejwt~=5.3.1